refactor: move the error component to the components folder and create your style file
